+++ packages/carbon-graphs/tests/unit/controls/Graph/GraphPanning-spec.js
@@ -164,7 +164,7 @@ describe('Graph - Panning', () => {
             'transform',
           ),
         ).translate[0],
-      ).toBeCloserTo(50);
+      ).toBeCloserTo(68);
       const panData = {
         key: 'uid_1',
         values: [{
@@ -296,7 +296,7 @@ describe('Graph - Panning', () => {
       const regionElementAfterPanning = fetchElementByClass(styles.contentContainer);
       triggerEvent(window, 'resize', () => {
         expect(regionElementAfterPanning.getAttribute('x'))
-          .toBeGreaterThan(containerX);
+          .toEqual(containerX);
         done();
       });
     });